Refactor content registration into their own files   * Add: Create bootstrap folder in components for registration.   * Fix: Deleted st2api model.
