+++ st2api/st2api/cmd/api.py
@@ -8,9 +8,11 @@ from eventlet import wsgi
 from st2common import log as logging
 from st2common.models.db import db_setup
 from st2common.models.db import db_teardown
+import st2actions.bootstrap.registrar as actions_registrar
+import st2actionrunner.bootstrap.registrar as runner_registrar
+import st2reactor.bootstrap.registrar as rules_registrar
 from st2api import config
 from st2api import app
-from st2api import model
 
 
 eventlet.monkey_patch(
@@ -53,19 +55,19 @@ def __setup():
     # 5. register runnertypes and actions. The order is important because actions require action
     #    types to be present in the system.
     try:
-        model.register_runner_types()
+        runner_registrar.register_runner_types()
     except Exception as e:
         LOG.warning('Failed to register action types: %s', e, exc_info=True)
         LOG.warning('Not registering stock actions.')
     else:
         try:
-            model.register_actions()
+            actions_registrar.register_actions()
         except Exception as e:
             LOG.warning('Failed to register actions: %s', e, exc_info=True)
 
     # 6. register rules
     try:
-        model.register_rules()
+        rules_registrar.register_rules()
     except Exception as e:
         LOG.warning('Failed to register rules: %s', e, exc_info=True)
