+++ st2api/tests/base.py
@@ -1,8 +1,7 @@
 import tests.config
 from pecan.testing import load_test_app
 from oslo.config import cfg
-
-from st2api import model
+import st2actionrunner.bootstrap.registrar as runner_registrar
 from st2tests import DbTestCase
 
 
@@ -26,7 +25,7 @@ class FunctionalTest(DbTestCase):
 
         # TODO(manas) : register action types here for now. RunnerType registration can be moved
         # to posting to /runnertypes but that implies implementing POST.
-        model.register_runner_types()
+        runner_registrar.register_runner_types()
 
         cls.app = load_test_app(config=cfg_dict)
